fix warning again (for centos7)

+++ bctoolbox/src/tester.c
@@ -310,7 +310,7 @@ static void test_complete_message_handler(const CU_pTest pTest, const CU_pSuite 
 	 */
 	if (max_vm_kb) {
 		struct mallinfo minfo = mallinfo();
-		if (minfo.uordblks > max_vm_kb * 1024) {
+		if ((size_t)minfo.uordblks > max_vm_kb * 1024) {
 			bc_tester_printf(
 				bc_printf_verbosity_error,
 				"The program exceeded the maximum amount of memory allocatable (%i bytes), aborting now.\n",
@@ -526,7 +526,7 @@ void bc_tester_init(void (*ftester_printf)(int level, const char *format, va_lis
 
 void bc_tester_set_max_vm(long amax_vm_kb) {
 #ifdef __linux
-	max_vm_kb = amax_vm_kb;
+	max_vm_kb = (size_t)amax_vm_kb;
 	bc_tester_printf(bc_printf_verbosity_info, "Maximum virtual memory space set to %li kilo bytes", max_vm_kb);
 #else
 	bc_tester_printf(bc_printf_verbosity_error, "Maximum virtual memory space setting is only implemented on Linux.");
